refactor: Move protocol-agnostic interface to omaha/protocol package
All Omaha version-agnostic bits and pieces should be placed in
`omaha/protocol` package. Each Omaha version implementation is now
responsible for version string validation.
