+++ omaha/factory.go
@@ -1,10 +1,8 @@
 package omaha
 
 import (
-	"fmt"
-
+	"github.com/brave/go-update/omaha/protocol"
 	v3impl "github.com/brave/go-update/omaha/v3"
-	"github.com/brave/go-update/protocol"
 )
 
 // Factory creates protocol handlers for a specific version
@@ -18,10 +16,6 @@ type DefaultFactory struct{}
 
 // CreateProtocol returns a Protocol implementation for the requested version
 func (f *DefaultFactory) CreateProtocol(version string) (protocol.Protocol, error) {
-	if !IsProtocolVersionSupported(version) {
-		return nil, fmt.Errorf("unsupported protocol version: %s", version)
-	}
-
-	// Currently, all supported versions are Omaha v3.x
+	// Each version package now handles its own validation
 	return v3impl.NewProtocol(version)
 }
