--- omaha/utils.go
@@ -1,6 +1,0 @@
-package omaha
-
-// IsJSONRequest is used to check if JSON parser should be used
-func IsJSONRequest(contentType string) bool {
-	return contentType == "application/json"
-}
